Add SNMP traps support
From now on, if configured, a new thread will be spawn to monitor for
SNMP traps. At this moment, they will be sent in the same configured
topic:

- In v1 traps, monitor name will be trap oid, and monitor value will be
1
- In v2c traps, monitor name will be trap oid oid variable, and value
will be first not standard trap oid or timestamp variable in PDU.

In order to monitor for traps in a specific port, you have to use new
JSON config key {"conf":"snmp_traps":{"server_name":<server_name>}}.

Included tests are:
- v1 SNMP trap with no variables, 1 and 2 variables, and with enterprise
  oids (end and no end with 0)
- v2 SNMP trap with no variables, 1 and 2 variables, and ommitting
  timestamp and/or trap oid
- Left as TO-DO: Notification testing

To avoid race condition, test with traps will wait for monitor to print
"Listening for traps on ...". In other case, SNMP trap would be sent
before listening server setup, and it will be lost.

As special case, IfUp and IfDown snmp interfaces ID will be transformed
in output JSON. By default, the trap format would be
`{...,"snmp_if_id.1": null}`. Monitor will output it as
`{...,"snmp_if_id": 1}`. In the future, we could do this transformation
to any SNMP trap variable.

As a minor change, monitor will not exit if no sensor to poll defined.
This way we can monitor only for traps, and configure for no polling.