FIX Join traps thread at exit
Due to not to join traps thread after pthread_cancel a lot of resources
were leaked from valgrind point of view. However, under musl 1.1.18, and
if we use a block select, thread is never cancelled.

In order to follow net-snmp code to block, we change the cancel type,
and we only allow traps thread to be cancelled at specific points.