Move docker logic from Makefile to Dockerfile
Previously, dockerfile was not able to build the full project: it
generated a development container with all the tools needed to build and
test the app, and a release container that needed the binary and tools
generated in the host and copied via build context.

Now Dockerfile contains many containers to generate:
- Previous development images has been splitted in two: mon-dev-base,
with all the tools needed to build project, and mon-dev, with extra
tools to be able to test the project pytest, valgrind, ...)
- Release container, that is now able to build the full project using
the previous container, with no need of manual intervention.

Also, previous m4 trickery was needed for generate both development and
release docker. Now it's not needed anymore.

Beyond that, .dockerfile was written using erroneously .gitignore
format. New fixed one avoid errors if the developer is trying to compile
the app in the build context, and try to build project. Docker build
will start from scratch and will must use git checkout'ed files.

More fixes:
- Use of dockerhub registry by default.
- Makefile's docker targets now added to .PHONY list.
- Cache and image build per branch per images shasum. Only save
  interesting containers.
- Clean of `.dockerignore`.